More byte conversion, adding Entrance links

+++ src/version_7/Coord2D.java
@@ -2,27 +2,27 @@ package version_7;
  
 public class Coord2D {
  
-    int x;
-    int y;
+    byte x;
+    byte y;
      
-    public Coord2D(int x, int y) {
+    public Coord2D(byte x, byte y) {
         this.x = x;
         this.y = y;
     }
  
-    public int getX() {
+    public byte getX() {
         return x;
     }
  
-    public void setX(int x) {
+    public void setX(byte x) {
         this.x = x;
     }
  
-    public int getY() {
+    public byte getY() {
         return y;
     }
  
-    public void setY(int y) {
+    public void setY(byte y) {
         this.y = y;
     }
